// libs/common/src/app/permission/permission.type.ts
export enum Action {
    Create = 'create',
    Read = 'read',
    Update = 'update',
    Delete = 'delete',
    Manage = 'manage',
}

export enum Subject {
    User = 'User',
    Workspace = 'Workspace',
    Workbook = 'Workbook',
    WorkspaceMember = 'WorkspaceMember',
    Book = 'Book',
    Note = 'Note',
}